File storage

+++ controllers/admin-controller.js
@@ -1,5 +1,6 @@
 const { validationResult } = require("express-validator");
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   return res.render("admin/add-edit-product", {
@@ -15,8 +16,31 @@ exports.getAddProduct = (req, res, next) => {
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
   const description = req.body.description;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
+
+  console.log(image);
+
+  if (!image) {
+    // image is not accepted or not set.
+    return res.status(422).render("admin/add-edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      errorMessage: "product image must be an image file",
+      validationErrors: [],
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+    });
+  }
+
+  const imageUrl = image.path;
+
+  console.log(image);
   const product = new Product({
     title: title,
     price: price,
@@ -37,7 +61,6 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array(),
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description,
       },
@@ -87,10 +110,11 @@ exports.postEditProduct = (req, res, next) => {
   const productId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedDescription = req.body.description;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedPrice = req.body.price;
 
   const errors = validationResult(req);
+
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/add-edit-product", {
       pageTitle: "Edit Product",
@@ -102,7 +126,6 @@ exports.postEditProduct = (req, res, next) => {
       product: {
         _id: productId,
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
       },
@@ -117,7 +140,12 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        // delete old image
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+
       return product.save().then(() => {
         // this executed after product saved successfully
         console.log("UPDATED PRODUCT!");
@@ -150,8 +178,16 @@ exports.getAdminProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.params.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
-    // executed after deletion
+
+  // get the file by id
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("No Product was found!"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
     .then((result) => {
       console.log("Product Deleted!");
       res.redirect("/admin/products");
